HID Implementation Final Sprint - Initial Commit
This commit is to stash work completed over the past couple of days. It
is extremely messy and almost 100% unfunctional. Expect hourly change
commits as progress occurs over the next two days.

At this point, hardware configuration is effectively locked and can not
be changed as a physical prototype is almost constructed.